before the major refactoring

+++ cli/src/factory/add_slide_com.cpp
@@ -39,7 +39,7 @@ void AddSlideCom::execute()
 
 void AddSlideCom::process_args(const std::vector<std::string>& tokens)
 {
-    for (int i = 0; i < tokens.size(); ++i)
+    for (size_t i = 0; i < tokens.size(); ++i)
     {
         if (_options.find(tokens[i]) != _options.end())
         {
